fix(opencode): isolate plugin entry exports
